% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_profile.R
\name{bs_profile}
\alias{bs_profile}
\title{List repositories}
\usage{
bs_profile(target, ...)
}
\arguments{
\item{target}{(character) Internal name of a single repository as
delivered in \code{\link[=bs_repositories]{bs_repositories()}}}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
a data.frame, of two columns: "name", "value".
"name" holds the "value" description. you can pivot the
data.frame to wide by e.g., \code{tidyr::pivot_wider(x)}
}
\description{
List repositories
}
\examples{
\dontrun{
res <- bs_repositories(coll = "de")
bs_profile(target = res$internal_name[1])
bs_profile(target = res$internal_name[2])
bs_profile(target = res$internal_name[20])
bs_profile(target = res$internal_name[110])
}
}
