% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_repositories.R
\name{bs_repositories}
\alias{bs_repositories}
\title{List repositories for a collection}
\usage{
bs_repositories(coll, ...)
}
\arguments{
\item{coll}{(character) collection code. For existing, pre-defined
collections see Appendix, section "Collection-related queries" in
the Appendix of
https://www.base-search.net/about/download/base_interface.pdf}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
a data.frame of two columns: "name", "internal_name"
}
\description{
List repositories for a collection
}
\examples{
\dontrun{
res <- bs_repositories(coll = "ceu")
bs_repositories(coll = "ceu")
bs_repositories(coll = "denw")
bs_repositories(coll = "de")
}
}
