% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_search.R
\name{bs_retry_options}
\alias{bs_retry_options}
\title{bs_search retry options}
\usage{
bs_retry_options(
  pause_base = 1,
  pause_cap = 60,
  pause_min = 1,
  times = 3,
  terminate_on = NULL,
  retry_only_on = NULL,
  onwait = NULL
)
}
\arguments{
\item{pause_base, pause_cap, pause_min}{basis, maximum, and minimum for
calculating wait time for retry.}

\item{times}{the maximum number of times to retry.}

\item{terminate_on, retry_only_on}{a vector of HTTP status codes.}

\item{onwait}{a callback function if the request will be retried and
a wait time is being applied.}
}
\value{
a named list with the parameters given to this function
}
\description{
bs_search retry options
}
\details{
see \link[crul:HttpClient]{crul::HttpClient} for more detailed explanation of these
parameters
}
