% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbace-package.R
\docType{package}
\name{rbace-package}
\alias{rbace-package}
\alias{rbace}
\title{rbace - Bielefeld Academic Search Engine Client}
\description{
rbace - Bielefeld Academic Search Engine Client
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
