% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbace-package.R
\docType{package}
\name{rbace-package}
\alias{rbace-package}
\alias{rbace}
\title{rbace - Bielefeld Academic Search Engine Client}
\description{
BASE API docs: https://www.base-search.net/about/download/base_interface.pdf
}
\details{
You may have to ask for access for one or more IP addresses you'll be
accessing BASE from, but you may not. It used to be the case, but now seems
like it's not IP restricted anymore. If you aren't able to access this API
you can request access at https://www.base-search.net/about/en/contact.php
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
